# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings, optionally within a date range and/or for one client
      description: >-
        Powers the calendar view (FR-10) — `start`/`end` (both `YYYY-MM-DD`, both
        optional, inclusive) filter on `date` rather than fetching every booking ever
        made. Omitting both returns every booking for the organization. `clientId`
        (Section 6.2 of PRD_DetailHub_Signup_Copy_and_Packages_Hardening.md) further
        narrows the result to bookings for that client — it composes with (never
        replaces) the organization scope every request is already limited to, so a
        `clientId` belonging to a different organization returns an empty array
        rather than another organization's booking.
      parameters:
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching bookings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      description: >-
        `packageIds` and `employeeSplit` fully describe the booking's line items and
        employee revenue split in one call (a `booking_packages`/`employee_splits`
        join row is written per entry). `employeeSplit` may be empty (a booking can be
        created before an employee is assigned), but if non-empty every percentage
        must be in `(0, 100]` and the set must sum to exactly 100.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "400":
          description: >-
            Invalid request body, an invalid `employeeSplit` (see this operation's
            description), or a `clientId`/`packageIds`/`employeeSplit[].employeeId`
            that doesn't reference a real row in this organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: The booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No booking with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      description: >-
        Partial update. Omitting `packageIds`/`employeeSplit` leaves them unchanged;
        passing either replaces the entire set (not a merge). Same `employeeSplit`
        validation as `POST /bookings`.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
      responses:
        "200":
          description: Updated booking.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "400":
          description: Invalid request body, an invalid `employeeSplit`, or a dangling reference.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No booking with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      description: >-
        Hard delete (unlike clients/packages/employees) — nothing else references a
        booking except its own `booking_packages`/`employee_splits` rows, which cascade.
        Prefer `PATCH /bookings/{id}` with `status: cancelled` for a normal
        cancellation; this is for genuine removal.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Booking deleted.
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No booking with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/payment:
    post:
      operationId: recordBookingPayment
      tags: [bookings]
      summary: Record how a booking was paid
      description: >-
        PRD_DetailHub_Payment_Methods.md FR-1/FR-2/FR-5. Selecting `zelle`/`venmo`/
        `cash` marks the booking paid via that method, timestamps it
        (`paymentRecordedAt`), and stores an optional free-text `paymentReference`.
        Selecting `credit_card` ALWAYS fails with 422 `card_processing_not_available`
        — no payment processor is connected in this codebase; this never charges a
        card, real or simulated.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordBookingPaymentRequest"
      responses:
        "200":
          description: Booking updated with the recorded payment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No booking with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: >-
            `paymentMethod` was `credit_card`, which isn't available yet — no payment
            processor is connected (FR-5). Never charges a card, real or simulated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/refund:
    post:
      operationId: refundBooking
      tags: [bookings]
      summary: Record a refund against a booking
      description: >-
        FR-8/Section 6.2 of PRD_DetailHub_Payment_Methods.md. Always a manual reversal
        record — sets `refundStatus: completed` and stores an optional free-text
        `refundReference`. Never calls a real processor refund API for any tender
        type; the money movement (if any) happens outside the app, same as the
        original Zelle/Venmo/Cash payment itself.
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundBookingRequest"
      responses:
        "200":
          description: Booking updated with the recorded refund.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResult"
        "400":
          description: Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not signed in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Signed in, but the session has no active organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No booking with this id exists in the current organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    EmployeeSplit:
      type: object
      description: >-
        Matches `artifacts/detail-hub/src/lib/mock-data.ts`'s `EmployeeSplit`
        interface. `Booking.employeeIds` is folded into this (see `BookingResult`) —
        the split's `employeeId` list already implies assignment, so a separate raw
        id list would just be redundant data that could drift.
      properties:
        employeeId:
          type: integer
        percentage:
          type: number
          minimum: 0
          maximum: 100
      required:
        - employeeId
        - percentage
...
    BookingResult:
      type: object
      description: >-
        Field names/shape match `artifacts/detail-hub/src/lib/mock-data.ts`'s
        `Booking` interface (`date`, `startTime`, `address`, not the master PRD's
        `scheduledDate`/`scheduledStartTime`/`serviceAddress` — the mock interface is
        the source of truth per Section 7), plus `createdAt`/`updatedAt`/`createdBy`.
        `employeeIds` is derived from `employeeSplit` for convenience, not stored
        separately. (`gasMeterStatus`/`weatherSnapshot` were removed per FR-9 of
        PRD_DetailHub_Signup_Copy_and_Packages_Hardening.md — dead placeholder
        columns never populated by a real integration; FR-11 says they come back for
        real later, see Fuel_Gauge_PRD.md.) `paymentMethod`/`paymentReference`/
        `paymentRecordedAt`/`refundStatus`/`refundReference` match
        `PRD_DetailHub_Payment_Methods.md` Section 7 exactly (superseding the earlier,
        rougher `paymentMethod`/`paymentNote` shape this API had before that PRD was
        reviewed — see `../../lib/db/src/schema/bookings.ts` for why) — settable only
        via `POST /bookings/{id}/payment` and `POST /bookings/{id}/refund`, never
        through `POST`/`PATCH /bookings`.
      properties:
        id:
          type: integer
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
        employeeIds:
          type: array
          items:
            type: integer
          description: Derived from `employeeSplit` (every `employeeId` present in the split).
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
        date:
          type: string
          format: date
        startTime:
          type: string
          description: "24-hour HH:MM, e.g. \"09:00\"."
        address:
          type: string
        depositAmount:
          type: number
        parkingCost:
          type: number
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          nullable: true
          enum: [zelle, venmo, cash, credit_card]
        paymentReference:
          type: string
          nullable: true
        paymentRecordedAt:
          type: string
          format: date-time
          nullable: true
        refundStatus:
          type: string
          nullable: true
          enum: [none, requested, completed]
        refundReference:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: id of the admin/owner user who created this booking.
      required:
        - id
        - clientId
        - packageIds
        - employeeIds
        - employeeSplit
        - date
        - startTime
        - address
        - depositAmount
        - parkingCost
        - status
        - createdAt
        - updatedAt
        - createdBy
    CreateBookingRequest:
      type: object
      properties:
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
          default: []
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
          default: []
        date:
          type: string
          format: date
        startTime:
          type: string
        address:
          type: string
          minLength: 1
        depositAmount:
          type: number
          minimum: 0
        parkingCost:
          type: number
          minimum: 0
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
      required:
        - clientId
        - date
        - startTime
        - address
        - depositAmount
        - parkingCost
        - status
    UpdateBookingRequest:
      type: object
      description: >-
        Partial update — all fields optional. Omitting `packageIds`/`employeeSplit`
        leaves them unchanged; passing either replaces the entire set. Payment/refund
        fields are not editable here — see `POST /bookings/{id}/payment` and
        `POST /bookings/{id}/refund`.
      properties:
        clientId:
          type: integer
        packageIds:
          type: array
          items:
            type: integer
        employeeSplit:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSplit"
        date:
          type: string
          format: date
        startTime:
          type: string
        address:
          type: string
          minLength: 1
        depositAmount:
          type: number
          minimum: 0
        parkingCost:
          type: number
          minimum: 0
        status:
          type: string
          enum: [confirmed, pending, completed, cancelled, no-show]
        notes:
          type: string
          nullable: true
    RecordBookingPaymentRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [zelle, venmo, cash, credit_card]
        paymentReference:
          type: string
          nullable: true
          description: Confirmation note for manual methods (e.g. last 4 of a Zelle confirmation, "exact change").
      required:
        - paymentMethod
    RefundBookingRequest:
      type: object
      properties:
        refundReference:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable detail, when available.
      required:
        - error